+++ inst/examples/ex-cdt.R
@@ -1,0 +1,11 @@
+## Create a factor table
+x <- data.frame(
+  A = c("a", "b", "a"),
+  B = c("x", "y", "z")
+)
+
+## Complete disjunctive table
+cdt(x)
+
+## Burt table
+burt(x)